+++ examples/hello_aphrodite/hello_aphrodite.cpp
@@ -162,8 +162,7 @@ void hello_aphrodite::init()
             aph::BufferLoadInfo bufferLoadInfo{ .debugName = "cube::vertex_buffer",
                                                 .data = vertices.data(),
                                                 .createInfo = {
-                                                    .size =
-                                                        static_cast<uint32_t>(vertices.size() * sizeof(vertices[0])),
+                                                    .size = vertices.size() * sizeof(vertices[0]),
                                                     .usage = ::vk::BufferUsageFlagBits::eStorageBuffer |
                                                              ::vk::BufferUsageFlagBits::eVertexBuffer,
                                                     .domain = aph::MemoryDomain::Device,
@@ -177,7 +176,7 @@ void hello_aphrodite::init()
             aph::BufferLoadInfo bufferLoadInfo{ .debugName = "cube::index_buffer",
                                                 .data = indices.data(),
                                                 .createInfo = {
-                                                    .size = static_cast<uint32_t>(indices.size() * sizeof(indices[0])),
+                                                    .size = indices.size() * sizeof(indices[0]),
                                                     .usage = ::vk::BufferUsageFlagBits::eStorageBuffer |
                                                              ::vk::BufferUsageFlagBits::eIndexBuffer,
                                                     .domain = aph::MemoryDomain::Device,
